Add mock test structure
