+++ src/cleaners/__mocks__/fetchShows.js
@@ -1,0 +1,1 @@
+export const fetchShows = jest.fn();
